Consolidate table enums into a single file
