+++ src/CMakeLists.txt
@@ -149,18 +149,15 @@ if(CELERITAS_USE_ROOT)
   include(${ROOT_USE_FILE})
 
   root_generate_dictionary(CeleritasRootInterface
+    base/OpaqueId.hh
+    io/GdmlGeometryMap.hh
+    io/ImportElement.hh
+    io/ImportMaterial.hh
     io/ImportParticle.hh
     io/ImportPhysicsTable.hh
     io/ImportPhysicsVector.hh
-    io/ImportProcess.hh
-    io/ImportModel.hh
-    io/ImportTableType.hh
-    io/GdmlGeometryMap.hh
-    io/ImportMaterial.hh
-    io/ImportElement.hh
     io/ImportVolume.hh
     physics/base/ParticleMd.hh
-    base/OpaqueId.hh
     LINKDEF io/RootInterfaceLinkDef.h
   )
