+++ src/io/ImportPhysicsTable.hh
@@ -8,16 +8,114 @@
 #pragma once
 
 #include <vector>
-
-#include "ImportProcessType.hh"
-#include "ImportTableType.hh"
-#include "ImportProcess.hh"
-#include "ImportModel.hh"
 #include "physics/base/ParticleMd.hh"
 #include "ImportPhysicsVector.hh"
 
 namespace celeritas
 {
+//---------------------------------------------------------------------------//
+/*!
+ * Property being described by the physics table.
+ *
+ * In Geant4 this is a string value.
+ */
+enum class ImportTableType
+{
+    not_defined,
+    dedx,
+    ionisation,
+    range,
+    range_sec,
+    inverse_range,
+    lambda,
+    lambda_prim,
+    lambda_mod_1,
+    lambda_mod_2,
+    lambda_mod_3,
+    lambda_mod_4
+};
+//---------------------------------------------------------------------------//
+/*!
+ * Category of physics process.
+ *
+ * See Geant4's G4ProcessType.hh for the equivalent enum.
+ */
+enum class ImportProcessType
+{
+    not_defined,
+    transportation,
+    electromagnetic,
+    optical,
+    hadronic,
+    photolepton_hadron,
+    decay,
+    general,
+    parameterisation,
+    user_defined,
+    parallel,
+    phonon,
+    ucn
+};
+
+//---------------------------------------------------------------------------//
+/*!
+ * Enumerator for the available physics processes.
+ *
+ * This enum was created to safely access the many physics tables imported.
+ */
+enum class ImportProcess
+{
+    not_defined,
+    ion_ioni,
+    msc,
+    h_ioni,
+    h_brems,
+    h_pair_prod,
+    coulomb_scat,
+    e_ioni,
+    e_brem,
+    photoelectric,
+    compton,
+    conversion,
+    rayleigh,
+    annihilation,
+    mu_ioni,
+    mu_brems,
+    mu_pair_prod,
+    transportation //!< Not a physics process
+};
+
+//---------------------------------------------------------------------------//
+/*!
+ * Enumerator for the available physics models.
+ *
+ * This enum was created to safely access the many imported physics tables.
+ */
+enum class ImportModel
+{
+    not_defined,
+    bragg_ion,
+    bethe_bloch,
+    urban_msc,
+    icru_73_qo,
+    wentzel_VI_uni,
+    h_brem,
+    h_pair_prod,
+    e_coulomb_scattering,
+    bragg,
+    moller_bhabha,
+    e_brem_sb,
+    e_brem_lpm,
+    e_plus_to_gg,
+    livermore_photoelectric,
+    klein_nishina,
+    bethe_heitler_lpm,
+    livermore_rayleigh,
+    mu_bethe_bloch,
+    mu_brem,
+    mu_pair_prod
+};
+
 //---------------------------------------------------------------------------//
 /*!
  * Store physics tables.
