+++ src/io/ImportPhysicsVector.hh
@@ -8,12 +8,26 @@
 #pragma once
 
 #include <vector>
-
-#include "ImportPhysicsVectorType.hh"
 #include "base/Types.hh"
 
 namespace celeritas
 {
+//---------------------------------------------------------------------------//
+/*!
+ * Geant4 equivalent enum for Physics vector types.
+ * [See Geant4's G4PhysicsVectorType.hh]
+ */
+enum class ImportPhysicsVectorType
+{
+    base,
+    linear,
+    log,
+    ln,
+    free,
+    ordered_free,
+    low_energy_free
+};
+
 //---------------------------------------------------------------------------//
 /*!
  * Store imported physics vector data [see Geant4's G4PhysicsVector.hh].
