--- src/io/ImportProcess.hh
@@ -1,46 +1,0 @@
-//----------------------------------*-C++-*----------------------------------//
-// Copyright 2020 UT-Battelle, LLC, and other Celeritas developers.
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file ImportProcess.hh
-//---------------------------------------------------------------------------//
-#pragma once
-
-#include <string>
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-/*!
- * Enumerator for the available physics processes.
- *
- * This enum was created to safely access the many physics tables imported.
- */
-enum class ImportProcess
-{
-    // EM process list
-    not_defined,
-    ion_ioni,
-    msc,
-    h_ioni,
-    h_brems,
-    h_pair_prod,
-    coulomb_scat,
-    e_ioni,
-    e_brem,
-    photoelectric,
-    compton,
-    conversion,
-    rayleigh,
-    annihilation,
-    mu_ioni,
-    mu_brems,
-    mu_pair_prod,
-
-    // Extra
-    transportation
-};
-
-//---------------------------------------------------------------------------//
-} // namespace celeritas
