--- src/io/ImportProcessType.hh
@@ -1,36 +1,0 @@
-//----------------------------------*-C++-*----------------------------------//
-// Copyright 2020 UT-Battelle, LLC, and other Celeritas developers.
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file ImportProcessType.hh
-//---------------------------------------------------------------------------//
-#pragma once
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-/*!
- * Enumerator for physics process types.
- *
- * See Geant4's G4ProcessType.hh for the equivalent enum.
- */
-enum class ImportProcessType
-{
-    not_defined,
-    transportation,
-    electromagnetic,
-    optical,
-    hadronic,
-    photolepton_hadron,
-    decay,
-    general,
-    parameterisation,
-    user_defined,
-    parallel,
-    phonon,
-    ucn
-};
-
-//---------------------------------------------------------------------------//
-} // namespace celeritas
